🎨 extract common generic utils
