+++ include/ros/type_traits.hpp
@@ -4,6 +4,20 @@
 #include <ros/operations.hpp>
 
 namespace ros {
+// generic type traits
+namespace detail {
+template <typename T, typename enable = void> struct unwrap_enum {
+    using type = T;
+};
+
+template <typename T>
+struct unwrap_enum<T, typename std::enable_if_t<std::is_enum_v<T>>> {
+    using type = std::underlying_type_t<T>;
+};
+
+template <typename T> using unwrap_enum_t = typename unwrap_enum<T>::type;
+} // namespace detail
+
 // field type traits
 namespace detail {
